Move SelectedCourses out
