+++ my-app/src/App.module.css
@@ -4,7 +4,6 @@
 }
 
 .left {
-    width: 200px;
     min-height: 100vh;
     padding-top: 20px;
 }
